src/requirements.txt
sanic==22.9.1
sanic-ext==22.9.1
PyJWT==2.6.0
scikit-image==0.19.3
scikit-learn==1.1.3
texttable==1.6.7
orjson==3.8.3
colorama==0.4.6
huey==2.4.4
